Move "createDisplayData" into previewActionsRoute file
