+++ src/components/actions/helpers.ts
@@ -3,7 +3,6 @@ import { ActionInTemplate } from '../../types'
 import evaluateExpression from '@openmsupply/expression-evaluator'
 import { BasicObject, IParameters } from '@openmsupply/expression-evaluator/lib/types'
 import DBConnect from '../databaseConnect'
-import { ActionQueueStatus } from '../../generated/graphql'
 
 export async function evaluateParameters(
   parameterQueries: BasicObject,
@@ -49,57 +48,4 @@ export const swapOutAliasedAction = async (templateId: number, action: ActionInT
   aliasedAction.parameter_queries = merge(aliasedAction.parameter_queries, overrideParams)
 
   return aliasedAction
-}
-
-interface ActionResult {
-  action: string // code
-  status: ActionQueueStatus
-  output: BasicObject | null
-  errorLog: string | null
-}
-
-interface ActionResultDisplayData {
-  type: 'DOCUMENT' | 'NOTIFICATION' | 'OTHER'
-  status: ActionQueueStatus
-  displayString: string
-  text?: string // for email string
-  fileId?: string // for generating url in front-end
-  errorLog: string | null // if Action failed
-}
-
-// Convert the full action output into a simplified format that can be easily
-// displayed by the front-end Preview module
-export const createDisplayData = (actionsOutput: ActionResult[]): ActionResultDisplayData[] => {
-  return actionsOutput.map((result) => {
-    switch (result.action) {
-      case 'sendNotification':
-        return {
-          type: 'NOTIFICATION',
-          status: result.status,
-          displayString: result.output?.notification?.subject ?? 'Email notification',
-          text: result.output?.notification?.message,
-          errorLog: result.errorLog,
-        }
-      case 'generateDoc':
-        return {
-          type: 'DOCUMENT',
-          status: result.status,
-          displayString:
-            result.output?.document?.description ??
-            result.output?.document?.filename ??
-            'Generated Document',
-          fileId: result.output?.document?.uniqueId,
-          errorLog: result.errorLog,
-        }
-      // We're only expecting preview results from sendNotification and generateDoc actions. Fallback for others:
-      default:
-        return {
-          type: 'OTHER',
-          status: result.status,
-          displayString: `Output of action: ${result.action}`,
-          text: JSON.stringify(result.output, null, 2),
-          errorLog: result.errorLog,
-        }
-    }
-  })
 }
